refactor: move opt-in methods to partial_set_security.go
