+++ x/ccv/provider/keeper/permissionless.go
@@ -176,115 +176,3 @@ func (k Keeper) IsConsumerActive(ctx sdk.Context, consumerId string) bool {
 		phase == types.CONSUMER_PHASE_INITIALIZED ||
 		phase == types.CONSUMER_PHASE_LAUNCHED
 }
-
-// GetOptedInConsumerIds returns all the consumer ids where the given validator is opted in
-func (k Keeper) GetOptedInConsumerIds(ctx sdk.Context, providerAddr types.ProviderConsAddress) (types.ConsumerIds, error) {
-	store := ctx.KVStore(k.storeKey)
-	bz := store.Get(types.ProviderConsAddrToOptedInConsumerIdsKey(providerAddr))
-	if bz == nil {
-		return types.ConsumerIds{}, nil
-	}
-
-	var consumerIds types.ConsumerIds
-	if err := consumerIds.Unmarshal(bz); err != nil {
-		return types.ConsumerIds{}, fmt.Errorf("failed to unmarshal consumer ids: %w", err)
-	}
-
-	return consumerIds, nil
-}
-
-// AppendOptedInConsumerId appends given consumer id to the list of consumers that validator has opted in
-// TODO (PERMISSIONLESS) -- combine it with SetOptedIn
-func (k Keeper) AppendOptedInConsumerId(ctx sdk.Context, providerAddr types.ProviderConsAddress, consumerId string) error {
-	store := ctx.KVStore(k.storeKey)
-
-	consumers, err := k.GetOptedInConsumerIds(ctx, providerAddr)
-	if err != nil {
-		return err
-	}
-
-	consumersWithAppend := types.ConsumerIds{
-		Ids: append(consumers.Ids, consumerId),
-	}
-
-	bz, err := consumersWithAppend.Marshal()
-	if err != nil {
-		return err
-	}
-
-	store.Set(types.ProviderConsAddrToOptedInConsumerIdsKey(providerAddr), bz)
-	return nil
-}
-
-// RemoveOptedInConsumerId removes the consumer id from this validator because it is not opted in anymore
-func (k Keeper) RemoveOptedInConsumerId(ctx sdk.Context, providerAddr types.ProviderConsAddress, consumerId string) error {
-	store := ctx.KVStore(k.storeKey)
-
-	consumers, err := k.GetOptedInConsumerIds(ctx, providerAddr)
-	if err != nil {
-		return err
-	}
-
-	if len(consumers.Ids) == 0 {
-		return fmt.Errorf("no consumer ids found for this provviderAddr: %s", providerAddr.String())
-	}
-
-	// find the index of the consumer we want to remove
-	index := -1
-	for i := 0; i < len(consumers.Ids); i++ {
-		if consumers.Ids[i] == consumerId {
-			index = i
-			break
-		}
-	}
-
-	if index == -1 {
-		return fmt.Errorf("failed to find consumer id (%s)", consumerId)
-	}
-
-	if len(consumers.Ids) == 1 {
-		store.Delete(types.ProviderConsAddrToOptedInConsumerIdsKey(providerAddr))
-		return nil
-	}
-
-	consumersWithRemoval := types.ConsumerIds{
-		Ids: append(consumers.Ids[:index], consumers.Ids[index+1:]...),
-	}
-
-	bz, err := consumersWithRemoval.Marshal()
-	if err != nil {
-		return err
-	}
-
-	store.Set(types.ProviderConsAddrToOptedInConsumerIdsKey(providerAddr), bz)
-	return nil
-}
-
-// IsValidatorOptedInToChainId checks if the validator with `providerAddr` is opted into the chain with the specified `chainId`.
-// It returns `found == true` and the corresponding chain's `consumerId` if the validator is opted in. Otherwise, it returns an empty string
-// for `consumerId` and `found == false`.
-func (k Keeper) IsValidatorOptedInToChainId(ctx sdk.Context, providerAddr types.ProviderConsAddress, chainId string) (string, bool) {
-	consumers, err := k.GetOptedInConsumerIds(ctx, providerAddr)
-	if err != nil {
-		k.Logger(ctx).Error("failed to retrieve the consumer ids this validator is opted in to",
-			"providerAddr", providerAddr,
-			"error", err)
-		return "", false
-	}
-
-	for _, consumerId := range consumers.Ids {
-		consumerChainId, err := k.GetConsumerChainId(ctx, consumerId)
-		if err != nil {
-			k.Logger(ctx).Error("cannot find chain id",
-				"consumerId", consumerId,
-				"error", err)
-			continue
-		}
-
-		if consumerChainId == chainId {
-			return consumerId, true
-		}
-
-	}
-	return "", false
-}
