+++ x/ccv/provider/keeper/permissionless_test.go
@@ -196,75 +196,3 @@ func TestConsumerPhase(t *testing.T) {
 	phase = providerKeeper.GetConsumerPhase(ctx, CONSUMER_ID)
 	require.Equal(t, providertypes.CONSUMER_PHASE_LAUNCHED, phase)
 }
-
-// TestOptedInConsumerIds tests the `GetOptedInConsumerIds`, `AppendOptedInConsumerId`, and `RemoveOptedInConsumerId` methods
-func TestGetOptedInConsumerIds(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	providerAddr := providertypes.NewProviderConsAddress([]byte("providerAddr"))
-	consumers, err := providerKeeper.GetOptedInConsumerIds(ctx, providerAddr)
-	require.NoError(t, err)
-	require.Empty(t, consumers)
-
-	err = providerKeeper.AppendOptedInConsumerId(ctx, providerAddr, "consumerId1")
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetOptedInConsumerIds(ctx, providerAddr)
-	require.NoError(t, err)
-	require.Equal(t, providertypes.ConsumerIds{
-		Ids: []string{"consumerId1"},
-	}, consumers)
-
-	err = providerKeeper.AppendOptedInConsumerId(ctx, providerAddr, "consumerId2")
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetOptedInConsumerIds(ctx, providerAddr)
-	require.NoError(t, err)
-	require.Equal(t, providertypes.ConsumerIds{
-		Ids: []string{"consumerId1", "consumerId2"},
-	}, consumers)
-
-	err = providerKeeper.AppendOptedInConsumerId(ctx, providerAddr, "consumerId3")
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetOptedInConsumerIds(ctx, providerAddr)
-	require.NoError(t, err)
-	require.Equal(t, providertypes.ConsumerIds{
-		Ids: []string{"consumerId1", "consumerId2", "consumerId3"},
-	}, consumers)
-
-	// remove all the consumer ids
-	err = providerKeeper.RemoveOptedInConsumerId(ctx, providerAddr, "consumerId2")
-	require.NoError(t, err)
-	consumers, err = providerKeeper.GetOptedInConsumerIds(ctx, providerAddr)
-	require.NoError(t, err)
-	require.Equal(t, providertypes.ConsumerIds{
-		Ids: []string{"consumerId1", "consumerId3"},
-	}, consumers)
-
-	err = providerKeeper.RemoveOptedInConsumerId(ctx, providerAddr, "consumerId3")
-	require.NoError(t, err)
-
-	err = providerKeeper.RemoveOptedInConsumerId(ctx, providerAddr, "consumerId1")
-	require.NoError(t, err)
-
-	consumers, err = providerKeeper.GetOptedInConsumerIds(ctx, providerAddr)
-	require.NoError(t, err)
-	require.Empty(t, consumers)
-}
-
-func TestIsValidatorOptedInToChain(t *testing.T) {
-	providerKeeper, ctx, ctrl, _ := testkeeper.GetProviderKeeperAndCtx(t, testkeeper.NewInMemKeeperParams(t))
-	defer ctrl.Finish()
-
-	chainId := "chainId"
-	providerAddr := providertypes.NewProviderConsAddress([]byte("providerAddr"))
-	_, found := providerKeeper.IsValidatorOptedInToChainId(ctx, providerAddr, chainId)
-	require.False(t, found)
-
-	expectedConsumerId := CONSUMER_ID
-	providerKeeper.SetConsumerChainId(ctx, expectedConsumerId, chainId)
-	providerKeeper.SetOptedIn(ctx, expectedConsumerId, providerAddr)
-	providerKeeper.AppendOptedInConsumerId(ctx, providerAddr, expectedConsumerId)
-	actualConsumerId, found := providerKeeper.IsValidatorOptedInToChainId(ctx, providerAddr, chainId)
-	require.True(t, found)
-	require.Equal(t, expectedConsumerId, actualConsumerId)
-}
